Expand Formulae for a Hypothesis

This is an internal function that helps to create the appropriate formula combinations for adding hypotheses.

expand_formula(formula, combination, table = FALSE, ...)

Arguments

formula
Formula showing relationship of outcomes and predictors, and is essentially the hypothesis. It allows for complex formulae e.g. multiple predictors, multiple exposures, and multiple covariates, that can be organized into individual formulae based on additional modifiers. Please see the for further documentation.

combination
The building pattern for how to put together the overall plan. It defines variable relationships that will be used. The options of combination will lead to the following patterns:

direct will not modify the original equation pattern

$$y ~ x1 + x2$$

sequential will create a list of sequentially adjusted formula

$$y ~ x1$$ $$y ~ x1 + x2$$

parallel will create a list of formula with parallel predictors

$$y ~ x1$$ $$y ~ x2$$

table
Determines if the returned value should be a list of formulas or a table. Defaults to FALSE, leading to a list generation.

...
Not currently used

Value

List or table of formulae and extracted terms (to be determined by the table argument).

Formulae

When supplying a formula object, each RHS term of the formula is considered an outcome variable, and is analyzed as a single outcome. Each LHS term is consider a predictor, and can be modified as below:

X() is placed around a term to define as an independent exposure, which will be placed in separate formulae from any other term marked as an exposure

F() is placed around a term for any predictors that should be maintained/fixed in all models, which can include complex terms, such as mixed effects

C() is placed around a term for any predictors that are a potential confounder, which does not yet change the relationship of how formulas are built, but is stored for future methods to be used with hypothesize()

For example, the equation below describes two independent exposures "x1" and "x2" that should be conditionally tested for every level of "z"

$$y ~ X(x1) + X(x2) + x3 + x4 + F((1 | z))$$