Set Components of a Hypothesis

The following functions will replace the internal attributes of a hypothesis object, without warning.

set_formulae() expands the formula based on outcomes and predictors (including both exposures, covariates, and fixed variables). The pattern is decided by the combination argument. The major functionality is within the expand_formula() function, but this allows for setting internal attributes.

set_data() adds data to a formula (internal to the hypothesize() function), along with potential stratification instructions. Modifies the formula by adding the attributes to the formula.

set_tests() Add the test and test options (if needed) the the hypothesis object. This modifies the attributes of the original object and invisibly returns the same object.

set_formulae(hypothesis, formula, combination)

set_data(hypothesis, data, data_name, strata = NULL)

set_tests(hypothesis, test, .test_opts)

set_vars(hypothesis, formula, confounders = NULL)

Arguments

hypothesis
A hypothesis object

combination
The pattern of how to generate formulas

data
A data frame object

data_name
Abstracted from the data argument

strata
Optional argument to subset data frame

Value

Invisibly returns the hypothesis object