Make a table of sequentially fit models using gt

This function is a wrapper for the gt package for quickly and easily making model tables. It is built for sequentially adjusted models that have been created from the extract_models() function, which is primarily uses the broom::tidy() function to help describe model fits.

tbl_sequential(
  x,
  var_col = "term",
  var_list,
  group_var = NULL,
  rank_col = "number",
  ranks,
  rank_lab = NULL,
  rank_footnotes = NULL,
  disp_col = c("estimate", "conf.low", "conf.high"),
  disp_glue = "{1} ({2}, {3})",
  stat_col = NA,
  stat_val = 0.05,
  stat_style = "fill",
  stat_opts = list(color = "lightgreen"),
  decimals = 2,
  ...
)

Arguments

x
A data table with columns that are similar to that from broom::tidy(), along with an naming variable that identifies which model the data is coming from. The default options presume that the data was generated from broom::tidy() with confidence intervals included.

var_col
Character name of the column containing the independent variable terms.

var_list
Named list of the primary exposure variables that were contained within each model, along with a text description (e.g. list(x = "primary_variable")).

group_var
Defaults to NULL. If present, presumes a grouping variable or level of analyses were present. Each group is given the label that represents the value of the level. To be more informative, can relabel the values prior to passing to the function.

rank_col
Character name of the column that serves as the ordering variable for the models, such as a sequence in the setting of sequential modeling.

ranks
Vector containing which models to select (either numerically or by a name), based on user preference. Serves as the base for the row names in the gt table. They are re-sequenced to be numerical integers if using the rank_lab parameter. For example, if the selections are ranks = c(1, 3, 5), then would become "rank_lab 1", "rank_lab 2", "rank_lab 3".

rank_lab
Defaults to NULL. Character string that would be prepended to each individual rank to provide legibility to the rows. For example, if ranks = c(1, 3, 5), then rank_lab = "Model" would produce the labels "Model 1", "Model 2", "Model 3". If missing, will not modify the row names.

rank_footnotes
Defaults to NULL. This is an unnamed list of the sequential labels to be added to help identify each model. The first argument is the label for the outcome variable. There then must be a value for each position selected by the ranks argument. For example, rank_footnotes = list("outcome", "exposure", "x1", c("x2", "x3")). This would generate a pattern:

\(Model 1 = outcome ~ exposure\)

\(Model 2 = outcome ~ Model 1 + x1\)

\(Model 3 = outcome ~ Model 2 + x2 + x3\)

disp_col
The numeric columns that contain the variables that should be displayed, such as disp_col = c("estimate", "conf.low", "conf.high"), which will be merged together in format chosen by the disp_glue argument.

disp_glue
A character vector that use the glue::glue() package to describe how the display columns should be arranged, such as disp_glue = "{1} ({2}, {3})". The numbers represent the corresponding position in the vector given to the disp_col argument (e.g. "1" represents "estimate").

stat_col
Occasionally the individual cells should be modified to show significance or importance of findings. This argument is a single character vector to name which column will help qualify that vector. For example, "p.value" may be used to help decide significance. Defaults to NA.

stat_val
If a statistic is chosen from the stat_col argument, then it is compared with and expected to be lower than the stat_val argument (stat_col < stat_val).

stat_style
If a statistic is chosen, then gt offers the gt::tab_style() function, which allows for modification. The options are c("fill", "text", "borders"), which correspond to cell_fill(), cell_text(), and cell_borders().

stat_opts
If a statistic is chosen, then a named list, such as list(color = "lightgreen") can be given to be used for the stat_style argument. There is not a validation method for this, so please read documentation from gt for this function.

decimals
The number of decimals to be used for displaying output. Defaults to 2 decimal places.

...
For passing additional arguments

See also